change: call the receive function in a differnt place
experimented by using the getCurrentPhase in the wait for green
function instead.
it seemed like the message queue wasn't getting cleared as expected, and rathern than try to clear it,
i thouht I'd try to call it at a different time.
However, in F5.B, it does explicity say to call the "receive" funciton